_projects/Data-Science/iiot-predictive-maintenance/requirements.txt
# Data Manipulation & Simulation
pandas==2.1.4
numpy==1.26.4
faker==22.0.0

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2
scipy==1.11.4

# Configuration & Quality
pyyaml==6.0.1
pydantic==2.6.0
pytest==8.0.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0